part 1 solution
* updates to day 1 to work with long instead of just int32
* fix bug in SumSearcher that caused it to not test some of the last values in the list

+++ aoc-2020/Day01/Day01.cs
@@ -10,9 +10,9 @@ namespace aoc2020
 		{
 			var target = 2020;
 			var values = File.ReadAllLines(@"Day01/input.txt")
-				.Select(line => int.Parse(line));
+				.Select(line => long.Parse(line));
 
-			var arr = new int[2] { 0, 0 };
+			var arr = new long[2] { 0, 0 };
 			var search = new SumSearcher(target, values.ToList());
 
 
@@ -24,7 +24,7 @@ namespace aoc2020
 				Console.WriteLine($"pt1 Sum {target} not found.");
 			}
 
-			arr = new int[3] { 0, 0, 0 };
+			arr = new long[3] { 0, 0, 0 };
 			if (search.FindSums(ref arr)) {
 				var product = arr[0];
 				for (int i = 1; i < arr.Length; i++) { product *= arr[i]; }
